requirements.txt
Flask
Flask-MySQLdb
groq
gunicorn
python-dotenv